+++ src/providers/autocomplete/index.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode';
 import { isCacheEnabled } from '../../configuration';
 import { cache } from '../cache';
 import { generateSuggestions } from './generate-suggestions';
-import { attributeRegex, extractInsertionData, applyRange } from './provider-helpers';
+import { attributeRegex, extractInsertionData } from './provider-helpers';
 
 export const autocompleteProviders = ['javascriptreact', 'typescriptreact']
   .map((language: string) => (
@@ -16,18 +16,18 @@ export const autocompleteProviders = ['javascriptreact', 'typescriptreact']
           return [];
         }
 
-        const { prefix, insertRange } = extractInsertionData(position, text, textContextMatch);
+        const { prefix, quoteMark, applyRange } = extractInsertionData(position, text, textContextMatch);
 
         if (isCacheEnabled() && cache[document.fileName]) {
-          return cache[document.fileName].map(applyRange(insertRange));
+          return cache[document.fileName].map(applyRange);
         }
 
-        return generateSuggestions(prefix, token).then(results => {
+        return generateSuggestions(prefix, quoteMark, token).then(results => {
           if (isCacheEnabled() && results.length > 0) {
             cache[document.fileName] = results;
           }
 
-          return results.map(applyRange(insertRange));
+          return results.map(applyRange);
         });
       },
     }, '/', "'", '"')