+++ src/providers/autocomplete/provider-helpers.ts
@@ -1,12 +1,10 @@
 import { Range, CompletionItem, Position } from 'vscode';
 import { attributes } from '../../configuration';
+import { isParamLink } from './generate-suggestions';
 
-export const attributeRegex = () => new RegExp(`(${attributes().join('|')})=(?<quotes>["'])(?<prefix>[^"']*)(?:["'])?`);
-
-export const applyRange = (range?: Range) => (item: CompletionItem) => ({
-  ...item,
-  range
-});
+export const attributeRegex = () => new RegExp(
+  `(${attributes().join('|')})=(?<quotes>"|'|\{\`)(?<prefix>[^"'\`]*)(?:["'\`])?`
+);
 
 export function extractInsertionData(position: Position, text: string, match: RegExpMatchArray) {
   const [_, attribute, quoteMark, prefix] = match ?? [];
@@ -24,5 +22,32 @@ export function extractInsertionData(position: Position, text: string, match: Re
     new Position(position.line, attrContent.end)
   );
 
-  return { prefix, insertRange };
+  return { prefix, quoteMark, applyRange: applyRange(quoteMark, insertRange) };
 }
+
+const applyRange = (quoteMark: string, range: Range) => (item: CompletionItem): CompletionItem => {
+  const newItem = { ...item, range };
+
+  if (!isParamLink(newItem.label.toString())) {
+    return newItem;
+  }
+
+  // adjust range to account for different quote marks
+  if (quoteMark !== '{`') {
+    newItem.additionalTextEdits = [{
+      newText: '{`',
+      range: new Range(
+        range.start.translate(0, -1),
+        range.start
+      )
+    }];
+    newItem.insertText += '`}';
+
+    newItem.range = new Range(
+      range.start,
+      range.end.translate(0, 1)
+    );
+  }
+
+  return newItem;
+};